+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.3] - 2024-12-08
+
+### Added
+- Context processor for site name and domain (`oxutils.context.site_name_processor`)
+  - Provides `site_name` and `site_domain` from settings to templates
+  - Easy integration with Django templates
+
 ## [0.1.0] - 2024-12-02
 
 ### Added
@@ -63,5 +70,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - django-auditlog, django-structlog, django-cid
 - PyJWT, jwcrypto, cryptography
 
-[Unreleased]: https://github.com/oxiliere/oxutils/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/oxiliere/oxutils/compare/v0.1.3...HEAD
+[0.1.3]: https://github.com/oxiliere/oxutils/compare/v0.1.0...v0.1.3
 [0.1.0]: https://github.com/oxiliere/oxutils/releases/tag/v0.1.0